begin using uint16_t

+++ code/include/trilobite/xtest.h
@@ -114,10 +114,10 @@ typedef struct {
  * passed, failed, ignored tests, the number of expected outcomes, and the number of assertions.
  */
 typedef struct {
-    int passed_count;     // Number of passed tests
-    int failed_count;     // Number of failed tests
-    int ignored_count;    // Number of ignored tests
-    int total_count;      // Total number of tests
+    uint16_t passed_count;     // Number of passed tests
+    uint16_t failed_count;     // Number of failed tests
+    uint16_t ignored_count;    // Number of ignored tests
+    uint16_t total_count;      // Total number of tests
 } XTestStats;
 
 /**
